+++ Changelog.md
@@ -1,7 +1,7 @@
 0.1.2 / TBD 
 ==================
-  * Network Learning 
-  	- out-of-graph learning
+  * Wrapped NetworkLearning.jl
+  * Bugfixes
 
 0.1.1 / 2017-10-27 
 ==================